# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/games": {
            "get": {
                "description": "Возвращает упорядоченный по id список игр с поддержкой limit/offset.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Получить список игр",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное число игр",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список игр и мета",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListGamesResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_mainpage.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_mainpage.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут обработки запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_mainpage.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт запись о новой игре.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Создание игры",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Игра успешно создана",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateGameResponse"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "UUID созданного ресурса (/games/{id})"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос (невалидный UUID, отсутствие полей, неверный формат даты)",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_creategametopic.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Конфликт — игра с таким именем уже существует",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_creategametopic.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_creategametopic.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут обработки запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_creategametopic.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/games/{game_id}/comments": {
            "get": {
                "description": "Возвращает упорядоченный по убыванию даты список комментариев к игре.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Получить список комментариев",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игры",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное число комментариев",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Сдвиг для пагинации",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список комментариев и мета",
                        "schema": {
                            "$ref": "#/definitions/handlers.ListCommentsResponse"
                        }
                    },
                    "400": {
                        "description": "Неверные параметры запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут обработки запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_listcomments.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет комментарий пользователя к указанной игре.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Постинг комментария",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID игры",
                        "name": "game_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса с полем user_id и text",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PostCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID созданного комментария",
                        "schema": {
                            "$ref": "#/definitions/handlers.AddCommentResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Игра не найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    },
                    "504": {
                        "description": "Таймаут запроса",
                        "schema": {
                            "$ref": "#/definitions/internal_controller_http_handlers_addcomment.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Comment": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "game_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.GameInList": {
            "type": "object",
            "properties": {
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
...
        "handlers.AddCommentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateGameRequest": {
            "type": "object",
            "properties": {
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "release_date": {
                    "description": "or time.Time + правильный UnmarshalJSON",
                    "type": "string"
                }
            }
        },
        "handlers.CreateGameResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ListCommentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Comment"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/internal_controller_http_handlers_listcomments.Pagination"
                }
            }
        },
        "handlers.ListGamesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.GameInList"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/internal_controller_http_handlers_mainpage.Pagination"
                }
            }
        },
        "handlers.PostCommentRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_controller_http_handlers_addcomment.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "машинно-читаемый код ошибки",
                    "type": "string"
                },
                "message": {
                    "description": "человеко-читаемое сообщение",
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_addcomment.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_addcomment.APIError"
                }
            }
        },
        "internal_controller_http_handlers_creategametopic.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "машинно-читаемый код ошибки",
                    "type": "string"
                },
                "message": {
                    "description": "человеко-читаемое сообщение",
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_creategametopic.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_creategametopic.APIError"
                }
            }
        },
...
        "internal_controller_http_handlers_listcomments.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_listcomments.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_listcomments.APIError"
                }
            }
        },
        "internal_controller_http_handlers_listcomments.Pagination": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_controller_http_handlers_mainpage.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_controller_http_handlers_mainpage.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_controller_http_handlers_mainpage.APIError"
                }
            }
        },
        "internal_controller_http_handlers_mainpage.Pagination": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}